requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yfinance>=0.2.31
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
